# mac80211: build each variant in its own dir; speed up build