If you compile both variants of mac80211 (regular & smallbuffers), the smallbuffers version will use the same build directory as the previously built regular version without running a make clean.  This causes the smallbuffers to use the same binaries as the regular version.  This only happens if CONFIG_AUTOREMOVE is unset.  So the packages built with the bots are not affected.

While at it, save considerable CPU time by only compiling the ath10k package when builiding the smallbuffers variant.  Here's the comparison when building with `CONFIG_ALL`

```
 make[1] package/mac80211/compile
 make[2] -C package/kernel/mac80211 compile
 make[2] -C package/kernel/mac80211 compile

real    2m50,515s
user    31m20,373s
sys     4m33,440s
```
```
 make[1] package/mac80211/compile
 make[2] -C package/kernel/mac80211 compile
 make[2] -C package/kernel/mac80211 compile

real    1m59,859s
user    19m19,130s
sys     2m49,541s
```